SizeFormの整形
